API: The GraphIndexes submodule

The GraphIndexes submodule contains some simple indexes and data structures that are useful if not required, for the development of mappers and motif/sequence finding heuristics.

All types and methods defined in this module are documented here.

Note

This is a reference of an internal sub-module's API for developers and experienced users. First ask yourself if what you need isn't covered by the higher-level WorkSpace API.

Warning

This submodule is still under construction.

Types

GenomeGraphs.GraphIndexes.UniqueMerIndex — Type

An index of all the unique kmers in a graph.

Public / Safe methods

Missing docstring.

Missing docstring for isunmappable. Check Documenter's build log for details.

Missing docstring.

Missing docstring for find_unique_mer. Check Documenter's build log for details.

Internal / Unsafe methods