API: The MerTools submodule

Note

This is a reference of an internal sub-module's API for developers and experienced users. First ask yourself if what you need isn't covered by the higher-level WorkSpace API.

Types

GenomeGraphs.MerTools.MerCount — Type

A simple mer count struct.

MerCount is a simple struct that binds a mer value to a count of the number of times it has been observed. This type, (sorted) vectors of them, and some additional utility methods, form the basic building blocks of the higher-level mer counting functionality of the MerTools sub-module.

Note

The count is stored as an UInt8 because often once the count is more than 255 we hardly care anymore.

GenomeGraphs.MerTools.MerCountHist — Type

A type for storing a frequency histogram of MerCounts, also sometimes referred to as a Kmer spectra.

GenomeGraphs.MerTools.DNAMerCount — Type

Shorthand for MerCount{DNAMer{K}}

GenomeGraphs.MerTools.RNAMerCount — Type

Shorthand for MerCount{RNAMer{K}}

Public / Safe methods

GenomeGraphs.MerTools.mer — Function

Get the mer from a MerCount.

GenomeGraphs.MerTools.freq — Function

Get the count from a MerCount.

Get the count from a MerCount, and convert it to type R.

GenomeGraphs.MerTools.collapse_into_counts — Function

collapse_into_counts(mers::Vector{M}) where {M<:AbstractMer}

Build a vector of sorted MerCounts from a Vector of a mer type.

This is a basic kernel function used for any higher level and more complex kmer counting procedures.

GenomeGraphs.MerTools.collapse_into_counts! — Function

collapse_into_counts!(result::Vector{MerCount{M}}, mers::Vector{M}) where {M<:AbstractMer}

Build a vector of sorted MerCounts from a Vector of a mer type.

This is a basic kernel function used for any higher level and more complex kmer counting procedures.

This is like collapse_into_counts, except it's first argument is a result vector that is cleared and filled with the result.

Note

The input vector mers will be sorted by this method.

GenomeGraphs.MerTools.merge_into! — Function

merge_into!(a::Vector{MerCount{M}}, b::Vector{MerCount{M}}) where {M<:AbstractMer}

Merge the MerCounts from vector b into the vector a.

Note

This will sort the input vectors a and b.

Internal / Unsafe methods

GenomeGraphs.MerTools.unsafe_collapse_into_counts! — Function

unsafe_collapse_into_counts!(result::Vector{MerCount{M}}, mers::Vector{M}) where {M<:AbstractMer}

Warning

This method is marked as unsafe because it assumes that the mers input vector is already sorted.

GenomeGraphs.MerTools.unsafe_merge_into! — Function

unsafe_merge_into!(a::Vector{MerCount{M}}, b::Vector{MerCount{M}}) where {M<:AbstractMer}

Merge the MerCounts from vector b into the vector a.

Warning

This method is marked as unsafe as it assumes both of the input vectors a and b are already sorted.