API: Graph submodule

Types

GenomeGraphs.SequenceDistanceGraph — Type

The SequenceDistanceGraph is a representation of a genome assembly. Sequences are contained in nodes, and the distances are represented by links.

A singe node represents a sequence and its reverse complement. Every node has a correlative ID starting from 1. For every node X in the graph, the negative ID -X is mapped to the reverse complement of X. This mapping is virtual: Only one node is stored in the graph. This is because every node has an orientaton: Each node has a positive end (+), and a negative end (-). So when a node is accessed with (or traversed by entering) the positive end the node yields the stored sequence. Conversely, when a node is accessed with (or traversed by entering) the negative end the node yelds the reverse complement of the stored sequence. In this way the positive end can be thought of as the sequence start, and the negative end can be thought of as the sequence end.

A single distance between two sequences is represented as a single link. Every link connects two node ends and contains a distance (they take the form ([+, -]n1, [+, -]n2, [+, -]dist)). A link connects two node ends, and so the order of the signed nodes in the links does not change the link. If the distance in a link is negative, this represents an overlap between two sequences. These overlaps must be "perfect overlaps".

This SequenceDistanceGraph type is only intended to be interacted with directly by developers and people who know what they are doing. Most tasks an end user wants to do to manipulate or query a graph can be

GenomeGraphs.SDG — Type

Shorthand for SequenceDistanceGraph

GenomeGraphs.DistanceGraphLink — Type

Represents a single distance between two sequences in a SequenceDistanceGraph.

Missing docstring.

Missing docstring for SDGNode. Check Documenter's build log for details.

Public / Safe methods

Graph Nodes and sequences

GenomeGraphs.name — Function

Get the name of the graph. Defaults to the symbol :sdg.

GenomeGraphs.n_nodes — Function

Get the number of nodes in the sequence distance graph sg.

GenomeGraphs.sequence — Function

sequence(sg::SequenceDistanceGraph, n::NodeID)

Get the full sequence of a node in a sequence distance graph using its correlative node id n.

Note

sequence accepts a NodeID that can be positive or negative. Nodes represent stretches of sequence in a canonical orientation, if you ask for for the sequence of say the third node, the positive node id 3 (which denotes traversing the third node in the forward direction), gives you the canonical sequence. If you use the negative ID -3 (which denotes traversing the third node in the reverse direction), you will get the reverse complement of the node's canonical (forward) sequence.

Note

It is safe to modify the returned sequence without screwing up your graph, yet thanks to BioSequences.jl's copy on write system for LongSequences, data copying will only occur if nessecery. You get the best of both worlds.

Get the sequence of the underlying SequenceDistanceGraph node.

Graph Topology

GenomeGraphs.find_link — Function

find_link(sg::SequenceDistanceGraph, src::NodeID, dst::NodeID)

Find and return the link that exists between a source node and a destination node, using their correlative node ids.

In this instance, the IDs also denote the "ends" of a node, a link connects.

Recall that every node has an orientaton: Each node has a positive end (+), and a negative end (-). So when a node is accessed with (or traversed by entering) the positive end the node yields the stored sequence. Conversely, when a node is accessed with (or traversed by entering) the negative end the node yelds the reverse complement of the stored sequence.

Thus find_link(sg, -5, 1) means you want to find a link in the graph that allows you to exit the (-) end of node 5 (meaning you just traversed it in the canonical + –> - orientation), and enter the (+) end of node 1 (meaning you will also traverse node one in the canonical + –> - orientation).

By contrast find_link(sg, -5, -1) means you want to find a link in the graph that allows you to exit the (-) end of node 5 - again having just traverse it in the + –> - orientation, but want to enter node 1 through it's (-) end - and thus traversing node 1 in the + <– - orientation, yielding the reverse complement of the canonical sequence of node 1.

If a link connecting the two node ends is present, you get it. If not, you get nothing. So make sure to check the output.

Note

This method is safe and public, but not reeeeaaallly intended for the end user, as they have to worry about links between "ends" of a node. In contrast to the rest of this framework, which rather has the user thinking in node centric terms. e.g. "I am at node 5 which means I have traversed it forwards or I am at node -5, which means I have traversed node 5 backwards. From here I can visit node 1, -3, and 2, so I may go through nodes 1 and 2 forwards, and through 3 backwards. The methods get_next_nodes and get_previous_nodes are more user-friendly and follow this more node-centric mindset.

GenomeGraphs.forward_links — Function

forward_links(sg::SequenceDistanceGraph, n::NodeID)

Get a vector of the links that are ahread of you, as you traverse node n, continuing forward in you present direction of travel (n can be a positive or negative node id).

The node id can be positive or negative. For example if you use a positive ID, such as 5, this means you are traversing node 5 in the canonical orientation, and so you will get the links that allow you to leave node 5 in the canonical direction. If you used a negative ID such as -5, that means you are traversing node 5 in the non-canonical orientation, and so you will get the links that allow you to leave node 5 in the non-canonical direction.

GenomeGraphs.backward_links — Function

backward_links(sg::SequenceDistanceGraph, n::NodeID)

Get a vector of the links that are behind of you, as you traverse node n, continuing forward in you present direction of travel (n can be a positive or negative node id).

The node id can be positive or negative. For example if you use a positive ID, such as 5, this means you are traversing node 5 in the canonical orientation, and so you will get the links that would allow you to enter node 5 in the canonical direction. If you used a negative ID such as -5, that means you are traversing node 5 in the non-canonical orientation, and so you will get the links that would allow you to enter node 5 in the non-canonical direction.

GenomeGraphs.get_next_nodes — Function

get_next_nodes(sg::SequenceDistanceGraph, n::NodeID)

Get the nodes you may visit next as you exit node n, maintaining your current direction of travel (n can be a positive or negative node ID).

Note

The node id can be positive or negative. For example if you use a positive ID, such as 5, this means you are traversing node 5 in the canonical orientation, and so you will get the nodes you may visit as you leave node 5 in the canonical direction. If you use a negative ID, such as -5, this means you are traversing node 5 in the non-canonical orientation, and so you will get the nodes you may visit as you leave node 5 in the non-canonical direction.

Note

The list of nodes returned are also signed to denote direction: Say you got [1, -2, 3] as a result of get_next_nodes(sg, 5). That means as you leave node 5 after traversing it in the canonical direction, you may proceed to travel through nodes 1 and 3 in the canonical direction, and node 2 in the non-canonical direction.

GenomeGraphs.get_previous_nodes — Function

get_previous_nodes(sg::SequenceDistanceGraph, n::NodeID)

Get the nodes you may have previously been on before you enterd node n, maintaining your current direction of travel (n can be a positive or negative node ID).

Note

The node id can be positive or negative. For example if you use a positive ID, such as 5, this means you are traversing node 5 in the canonical orientation, and so you will get the nodes you may have visited prior to entering node 5 in the canonical direction. If you use a negative ID, such as -5, this means you are traversing node 5 in the non-canonical orientation, and so you will get the nodes you may have visited prior to entering node 5 in the non-canonical direction.

Note

The list of nodes returned are also signed to denote direction: Say you got [1, -2, 3] as a result of get_previous_nodes(sg, 5). That means prior to entering node 5 in the canonical direction, you may have traveled through nodes 1 and 3 in the canonical direction, and node 2 in the non-canonical direction.

GenomeGraphs.find_tip_nodes — Function

find_tip_nodes(sg::SequenceDistanceGraph, min_size::Integer)

Get a set of IDs of all the nodes in the graph sg that count as "tips".

Here a tip node is defined as a node that only has a single neighbouring node at one of it's ends, and no neighbouring nodes out of one of its other end.

The node's sequence must also be larger than min_size base pairs in length.

GenomeGraphs.find_tip_nodes! — Function

find_tip_nodes!(result::Set{NodeID}, sg::SequenceDistanceGraph, min_size::Integer)

Get a set of IDs of all the nodes in the graph sg that count as "tips".

Here a tip node is defined as a node that only has a single neighbouring node at one of it's ends, and no neighbouring nodes out of one of its other end.

The node's sequence must also be larger than min_size base pairs in length.

Note

This method modifys a result input. So if you want to find tips in a graph repeatedly, you can resuse a Set, saving you some allocation costs.

Graph IO

Missing docstring.

Missing docstring for write_to_gfa1. Check Documenter's build log for details.

GenomeGraphs.load_from_gfa1! — Function

load_from_gfa1!(sg::SequenceDistanceGraph{S}, gfafile::AbstractString, fafile::AbstractString) where {S<:BioSequence}

Load a graph from a GFAv1 formatted file, and associated FASTA file of node sequences.

Note

The GFA format permits storing sequences of the graph nodes in a seperate fasta file, instead of in the GFA file. This is so as the sequences of the graph nodes can be easily fed into other tools that typically accept FASTA files as input. Many assemblers also output a GFA + FASTA combo. Therefore, this method asks for the filepath of a GFAv1 file, as well as a filepath to a FASTA formatted file. This method reads the node sequences from the FASTA file, before getting the links between nodes from the GFAv1 file.

Internal / Unsafe methods

Accessors and property queries

GenomeGraphs.check_node_id — Function

check_node_id(sg::SequenceDistanceGraph, i::NodeID)

The method responsible for checking that a node id used as input to a method that queries or edits a SequenceDistanceGraph is a sensible value.

GenomeGraphs.nodes — Function

Get a reference to the vector of nodes in a graph sg.

Warning

It is a bad idea to edit this vector yourself unless you know what you are doing.

GenomeGraphs.node_unsafe — Function

node_unsafe(sg::SequenceDistanceGraph, n::NodeID)

Get a refernece specific node from a sequence distance graph sg using its correlative node id n.

Note

node_unsafe accepts a NodeID that can be positive or negative. E.g. providing either 5 or -5 both mean node 5 in a graph, and so you will get the node for node 5.

Warning

This method returns a reference to a graph's underlying SDGNode. NOT a copy! Messing with this will screw up your graph. So this method is not recommended or exported - unless you really know what you're doing.

Warning

This method is explicitly marked unsafe for a reason. It does zero checking of the value it is passed as the node id. It also makes use of the @inbounds macro. This makes it faster, but you need to be 100% sure that your code won't try to call this with a bad id. So it is not exported or recommended

unless you really know what you're doing.

GenomeGraphs.sequence_unsafe — Function

sequence_unsafe(sg::SequenceDistanceGraph, n::NodeID)

Get the reference to a node's underlying sequence object.

Warning

This method is unsafe, no checking of node id's occurs and you get a reference to the node's sequence object - not a copy, so doing transformation operations (reverse_complement, setindex, etc.) to it will probably screw up the graph!

GenomeGraphs.links — Function

Get a reference to the vector of vectors of links in a graph sg.

Warning

It is a bad idea to edit this vector yourself unless you know what you are doing.

links(sg::SequenceGraph, n::NodeID)

Get a **reference** to a vector storing all the links of a node in a
`SequenceDistanceGraph`, the node is specified using its correlative node id `n`.

Note

links accepts a NodeID that can be positive or negative. E.g. providing either 5 or -5 both mean node 5 in a graph, and so you will get the links for node 5.

Warning

This method returns a reference to an underlying links vector that the SequenceDistanceGraph owns - NOT a copy! Messing with this will screw up your graph. So this method is not recommended or exported - unless you really know what you're doing.

GenomeGraphs.links_unsafe — Function

links_unsafe(sg::SequenceDistanceGraph, n::NodeID)

Get a reference to a vector storing all the links of a node in a SequenceDistanceGraph, the node is specified using its correlative node id n.

Note

links_unsafe accepts a NodeID that can be positive or negative. E.g. providing either 5 or -5 both mean node 5 in a graph, and so you will get the links for node 5.

Warning

This method returns a reference to an underlying links vector that the SequenceDistanceGraph owns - NOT a copy! Messing with this will screw up your graph. So this method is not recommended or exported - unless you really know what you're doing.

Warning

This method is explicitly marked unsafe for a reason. It does zero checking of the value it is passed as the node id. It also makes use of the @inbounds macro. This makes it faster, but you need to be 100% sure that your code won't try to call this with a bad id. So it is not exported or recommended

unless you really know what you're doing.