The Sequence Distance Graph (SDG)

GenomeGraphs.SequenceDistanceGraph — Type

The SequenceDistanceGraph is a representation of a genome assembly. Sequences are contained in nodes, and the distances are represented by links.

A singe node represents a sequence and its reverse complement. Every node has a correlative ID starting from 1. For every node X in the graph, the negative ID -X is mapped to the reverse complement of X. This mapping is virtual: Only one node is stored in the graph. This is because every node has an orientaton: Each node has a positive end (+), and a negative end (-). So when a node is accessed with (or traversed by entering) the positive end the node yields the stored sequence. Conversely, when a node is accessed with (or traversed by entering) the negative end the node yelds the reverse complement of the stored sequence. In this way the positive end can be thought of as the sequence start, and the negative end can be thought of as the sequence end.

A single distance between two sequences is represented as a single link. Every link connects two node ends and contains a distance (they take the form ([+, -]n1, [+, -]n2, [+, -]dist)). A link connects two node ends, and so the order of the signed nodes in the links does not change the link. If the distance in a link is negative, this represents an overlap between two sequences. These overlaps must be "perfect overlaps".

Querying an SDG for basic properties

GenomeGraphs.nodes — Function

Get a reference to the vector of nodes in a graph sg.

Warning

It is a bad idea to edit this vector yourself unless you know what you are doing.

GenomeGraphs.n_nodes — Function

Get the number of nodes in the sequence distance graph sg.

GenomeGraphs.each_node_id — Function

Iterate over every node ID in the sequence distance graph sg.

GenomeGraphs.links — Function

Get a reference to the vector of vectors of links in a graph sg.

Warning

It is a bad idea to edit this vector yourself unless you know what you are doing.

links(sg::SequenceGraph, n::NodeID)

Get all of the links of a Node of a sequence distance graph using its correlative node id n.

Note

links accepts a NodeID that can be positive or negative. E.g. providing either 5 or -5 both mean node 5 in a graph, and so you will get the links for node 5.

GenomeGraphs.node — Function

node(sg::SequenceDistanceGraph, n::NodeID)

Get a specific node from a sequence distance graph sg using its correlative node id n.

Note

node accepts a NodeID that can be positive or negative. E.g. providing either 5 or -5 both mean node 5 in a graph, and so you will get the links for node 5.

Get a view of node `nodeid` in the base SequenceDistanceGraph of the
WorkSpace `ws`.

GenomeGraphs.sequence — Function

sequence(sg::SequenceDistanceGraph, n::NodeID)

Get the full sequence of a node in a sequence distance graph using its correlative node id n.

Note

sequence accepts a NodeID that can be positive or negative. Nodes represent stretches of sequence in a canonical orientation, if you ask for for the sequence of say the third node, the positive node id 3 (which denotes traversing the third node in the forward direction), gives you the canonical sequence. If you use the negative ID -3 (which denotes traversing the third node in the reverse direction), you will get the reverse complement of the node's canonical (forward) sequence.

Note

It is safe to modify the returned sequence without screwing up your graph, yet thanks to BioSequences.jl's copy on write system for LongSequences, data copying will only occur if nessecery. You get the best of both worlds.

Get the sequence of the underlying SequenceDistanceGraph node.

Manually editing an SDG by manipulating nodes and links

It is not recommended you do this if you are a high level user. However these small editing operations are required for developers. If you find yourself needing these methods, you will have to explicitly import them, as they are not exported from the module.

If you find yourself wanting to edit the graph manually, it's a good idea to ask the package authors listed in the Project.toml or .github/CODEOWNERS.

Missing docstring.

Missing docstring for add_node!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_node!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_link!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_link!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disconnect_node!. Check Documenter's build log for details.

#Code example for testing the error correction and simplifying the graph

In the example below we generate a toy graph from 4-mers which contain a bubble. As for now, we make use of randomly generated coverage information for

using BioSequences
using BioSequenceGraphs

kmerlist = Vector{DNAKmer{4}}([DNAKmer{4}("ATAC"),DNAKmer{4}("TACG"),DNAKmer{4}("TACC"),DNAKmer{4}("ACGA"),DNAKmer{4}("CGAA"),DNAKmer{4}("GAAT"),DNAKmer{4}("AATC"),DNAKmer{4}("ACCA"),DNAKmer{4}("CCAA"),DNAKmer{4}("CAAT")])
sdg = SequenceDistanceGraph(kmerlist,true)

The new SequenceDistanceGraph constructor takes as input a list of kmerlist (not necessarily sorted or in canonical form) and a boolean for whether to do error correction or not. Then by using the coverage information generated randomly, simplifies the graph. The resulting SDG before and after the error correction steps are as follows:

***Before:***

SequenceDistanceGraph{BioSequence{DNAAlphabet{2}}}(SDGNode{BioSequence{DNAAlphabet{2}}}[SDGNode{BioSequence{DNAAlphabet{2}}}(AATC, false), SDGNode{BioSequence{DNAAlphabet{2}}}(ATAC, false), SDGNode{BioSequence{DNAAlphabet{2}}}(ATTCGTA, false), SDGNode{BioSequence{DNAAlphabet{2}}}(ATTGGTA, false)], Array{DistanceGraphLink,1}[[DistanceGraphLink(1, 3, -3), DistanceGraphLink(1, 4, -3)], [DistanceGraphLink(-2, -4, -3), DistanceGraphLink(-2, -3, -3)], [DistanceGraphLink(3, 1, -3), DistanceGraphLink(-3, -2, -3)], [DistanceGraphLink(4, 1, -3), DistanceGraphLink(-4, -2, -3)]])

***After:***

SequenceDistanceGraph{BioSequence{DNAAlphabet{2}}}(SDGNode{BioSequence{DNAAlphabet{2}}}[SDGNode{BioSequence{DNAAlphabet{2}}}(GATTCGTAC, false)], Array{DistanceGraphLink,1}[[]])

So the algorithm first generates a sdg with collapsing all simple paths using the initial kmer list. Then at the next stage the low covered bubble branch is removed and path collapsing is applied one more time. This results in a sdg with a whole single node, a perfect unitig.

##TODO:

Implement the sdg constructor to take as input a list of reads instead of kmers and generate all kmer + coverage information from them.