The Sequence Distance Graph (SDG)

The Sequence Distance Graph (SDG)

GenomeGraphs.SequenceDistanceGraph — Type.

The SequenceDistanceGraph is a representation of a genome assembly. Sequences are contained in nodes, and the distances are represented by links.

A singe node represents a sequence and its reverse complement. Every node has a correlative ID starting from 1. For every node X in the graph, the negative ID -X is mapped to the reverse complement of X. This mapping is virtual: Only one node is stored in the graph. This is because every node has an orientaton: Each node has a positive end (+), and a negative end (-). So when a node is accessed with (or traversed by entering) the positive end the node yields the stored sequence. Conversely, when a node is accessed with (or traversed by entering) the negative end the node yelds the reverse complement of the stored sequence. In this way the positive end can be thought of as the sequence start, and the negative end can be thought of as the sequence end.

A single distance between two sequences is represented as a single link. Every link connects two node ends and contains a distance (they take the form ([+, -]n1, [+, -]n2, [+, -]dist)). A link connects two node ends, and so the order of the signed nodes in the links does not change the link. If the distance in a link is negative, this represents an overlap between two sequences. These overlaps must be "perfect overlaps".

source

Querying an SDG for basic properties

GenomeGraphs.nodes — Function.

Get a reference to the vector of nodes in a graph sg.

Warning

It is a bad idea to edit this vector yourself unless you know what you are doing.

source

GenomeGraphs.n_nodes — Function.

Get the number of nodes in the sequence distance graph sg.

source

GenomeGraphs.each_node_id — Function.

Iterate over every node ID in the sequence distance graph sg.

source

GenomeGraphs.links — Function.

Get a reference to the vector of vectors of links in a graph sg.

Warning

It is a bad idea to edit this vector yourself unless you know what you are doing.

source

links(sg::SequenceGraph, n::NodeID)

Get all of the links of a Node of a sequence distance graph using its correlative node id n.

Note

links accepts a NodeID that can be positive or negative. E.g. providing either 5 or -5 both mean node 5 in a graph, and so you will get the links for node 5.

source

GenomeGraphs.node — Function.

node(sg::SequenceDistanceGraph, n::NodeID)

Get a specific node from a sequence distance graph sg using its correlative node id n.

Note

node accepts a NodeID that can be positive or negative. E.g. providing either 5 or -5 both mean node 5 in a graph, and so you will get the links for node 5.

source

GenomeGraphs.sequence — Function.

sequence(sg::SequenceDistanceGraph, n::NodeID)

Get the full sequence of a node in a sequence distance graph using its correlative node id n.

source

Manually editing an SDG by manipulating nodes and links

It is not recommended you do this if you are a high level user. However these small editing operations are required for developers. If you find yourself needing these methods, you will have to explicitly import them, as they are not exported from the module.

If you find yourself wanting to edit the graph manually, it's a good idea to ask the package authors listed in the Project.toml or .github/CODEOWNERS.

Missing docstring.

Missing docstring for add_node!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_node!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_link!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_link!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for disconnect_node!. Check Documenter's build log for details.